WIP - Allow zigbee to be PC comm usart
Allow user to configure the to use the zigbee pins as a tel interface
to talk to pc.

Currently is setup to hopefully detect when data is received on either
USB or Zigbee and define that as the USART to talk back to.

Builds - but has not been tested yet!